# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Angola Viva",
        "description": "Documentação da API do sistema Angola Viva",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
...
        "/viaturas": {
            "get": {
                "tags": [
                    "Viaturas"
                ],
                "summary": "Listar todas as viaturas",
                "operationId": "32562336e29d1230abafa34bbec5cde5",
                "responses": {
                    "200": {
                        "description": "Lista de viaturas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Viaturas"
                ],
                "summary": "Criar uma nova viatura",
                "operationId": "fc250467a3f6bbe82515994fe447da36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identificacao",
                                    "tipo",
                                    "status",
                                    "id_hospital"
                                ],
                                "properties": {
                                    "identificacao": {
                                        "type": "string"
                                    },
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "disponivel",
                                            "em_viagem",
                                            "ocupada"
                                        ]
                                    },
                                    "latitude": {
                                        "type": "number"
                                    },
                                    "longitude": {
                                        "type": "number"
                                    },
                                    "id_hospital": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Viatura criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/viaturas/{id}": {
            "get": {
                "tags": [
                    "Viaturas"
                ],
                "summary": "Exibir uma viatura",
                "operationId": "b635c45868d58aad41157a8bd082b815",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da viatura",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da viatura",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Viatura não encontrada"
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Viaturas"
                ],
                "summary": "Atualizar uma viatura",
                "operationId": "b594d28a76e77c8f4b4debfa8847c3ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da viatura",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "identificacao": {
                                        "type": "string"
                                    },
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "disponivel",
                                            "em_viagem",
                                            "ocupada"
                                        ]
                                    },
                                    "latitude": {
                                        "type": "number"
                                    },
                                    "longitude": {
                                        "type": "number"
                                    },
                                    "id_hospital": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Viatura atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Viatura não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Viaturas"
                ],
                "summary": "Remover uma viatura",
                "operationId": "dbbd8e2560c5888628a6beed059e90d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da viatura",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Viatura removida com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Viatura não encontrada"
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilize o token JWT no formato Bearer para autenticação.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}